Use Scala Pickling for serialization
The serialization mechanism of play2-memcached uses Java serialization
to handle the conversion of objects to binary and vice versa. This
system requires that objects to be serialised are of a type that
implements the Serializable interface. However, this breaks the contract
of the Play Cache API, as the default Cache API implementation, which
uses EHCache, does not have this constraint.

Using play2-memcached as the cache implementation causes plugins, such
as Play Authenticate to break, as they cache non-serializable objects
under the assumption that this is supported by the Cache API
implementation.

To allow the play2-memcached cache plugin to properly fulfill the
the Cache API contract, change the custom serialization class,
CustomSerializing, to use the Scala Pickling library for serialization,
and deserialization.